+++ src/Domain/HydraScript.Domain.FrontEnd/Parser/Impl/Ast/AbstractSyntaxTreeNode.cs
@@ -23,10 +23,10 @@ public abstract class AbstractSyntaxTreeNode : IAbstractSyntaxTreeNode
     protected virtual IReadOnlyList<IAbstractSyntaxTreeNode> Children { get; } = [];
 
     public IEnumerator<IAbstractSyntaxTreeNode> GetEnumerator() =>
-        Children.ToList().GetEnumerator();
+        Children.GetEnumerator();
 
     IEnumerator IEnumerable.GetEnumerator() =>
-        GetEnumerator();
+        Children.GetEnumerator();
 
     public int Count => Children.Count;
 
@@ -58,60 +58,4 @@ public abstract class AbstractSyntaxTreeNode : IAbstractSyntaxTreeNode
     protected abstract string NodeRepresentation();
     public override string ToString() =>
         $"{GetHashCode()} [label=\"{NodeRepresentation()}\"]";
-
-    public struct TraverseEnumerator : 
-        IEnumerator<IAbstractSyntaxTreeNode>, 
-        IEnumerable<IAbstractSyntaxTreeNode>
-    {
-        [ThreadStatic] private static Stack<IAbstractSyntaxTreeNode>? _buffer;
-
-        private IAbstractSyntaxTreeNode _current;
-        public IAbstractSyntaxTreeNode Current
-        {
-            [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            get => _current;
-        }
-
-        private readonly Stack<IAbstractSyntaxTreeNode> _stack;
-        public TraverseEnumerator(IAbstractSyntaxTreeNode parent)
-        {
-            var stack = _buffer ?? new Stack<IAbstractSyntaxTreeNode>(128);
-            _buffer = null;
-
-            foreach(var child in parent)
-                stack.Push(child);
-
-            _stack = stack;
-            _current = null!;
-        }
-        public bool MoveNext()
-        {
-            var stack = _stack;
-            if(stack.Count == 0)
-                return false;
-
-            var current = _stack.Pop();
-
-            foreach(var child in current)
-                stack.Push(child);
-
-            _current = current;
-            return true;
-        }
-
-        public void Dispose()
-        {
-            _stack.Clear();
-            _buffer = _stack;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public TraverseEnumerator GetEnumerator() => this;
-        public void Reset() { }
-        IEnumerator<IAbstractSyntaxTreeNode> IEnumerable<IAbstractSyntaxTreeNode>.GetEnumerator() => this;
-
-        IEnumerator IEnumerable.GetEnumerator() => this;
-
-        object IEnumerator.Current => Current;
-    }
 }
